native: render post plaintext preview from intermediate PostContent (#4752)
* [wip] Move post content utils to shared/

* Update imports for contentUtils.ts -> shared/logic/postContent.ts

* Add renderer from PostContent -> string

* Add optional plaintext preview to ChatMessage.All fixture

* Back `getTextContent` using plaintextPreviewOf (avoiding call changes)

* Use new getTextContent in existing callsites

* Delete unused code

* Use getTextContent in mobile notifications handling code

* Import necessary non-type members
