+++ packages/app/ui/components/AudioPost.tsx
@@ -1,3 +1,4 @@
+import { convertContent } from '@tloncorp/shared/logic';
 import { Pressable } from '@tloncorp/ui';
 import { useCallback, useMemo, useRef } from 'react';
 
@@ -5,7 +6,6 @@ import { RenderItemType } from '../contexts/componentsKits';
 import { ChatAuthorRow } from './AuthorRow';
 import { AudioPlayerHandle } from './Embed/AudioEmbedShared';
 import { AudioPlayer } from './Embed/AudioPlayer';
-import { convertContent } from './PostContent/contentUtils';
 
 export const AudioPost: RenderItemType = (props) => {
   const audioUrl = useMemo(() => {
