+++ packages/app/ui/components/PostContent/BlockRenderer.tsx
@@ -1,3 +1,4 @@
+import type * as cn from '@tloncorp/shared/logic';
 import { Image, Pressable, Text, useCopy } from '@tloncorp/ui';
 import { ImageLoadEventData } from 'expo-image';
 import React, {
@@ -21,11 +22,11 @@ import { VideoEmbed } from '../Embed';
 import EmbedContent from '../Embed/EmbedContent';
 import { HighlightedCode } from '../HighlightedCode';
 import { InlineRenderer } from './InlineRenderer';
-import * as cn from './contentUtils';
+import { ContentContext, useContentContext } from './contentUtils';
 
 export const BlockWrapper = styled(View, {
   name: 'ContentBlock',
-  context: cn.ContentContext,
+  context: ContentContext,
   padding: '$l',
   variants: {
     isNotice: {
@@ -121,7 +122,7 @@ function TextContent(props: ComponentProps<typeof LineText>) {
 export const LineText = styled(Text, {
   color: '$primaryText',
   size: '$body',
-  context: cn.ContentContext,
+  context: ContentContext,
   userSelect: 'text',
   cursor: 'text',
   variants: {
@@ -228,7 +229,7 @@ export function ImageBlock({
   block: cn.ImageBlockData;
   imageProps?: ComponentProps<typeof ContentImage>;
 } & ComponentProps<typeof View>) {
-  const { onPressImage, onLongPress } = cn.useContentContext();
+  const { onPressImage, onLongPress } = useContentContext();
   const [dimensions, setDimensions] = useState({
     width: block.width || null,
     height: block.height || null,
@@ -307,7 +308,7 @@ export function ImageBlock({
 
 const ContentImage = styled(Image, {
   name: 'ContentImage',
-  context: cn.ContentContext,
+  context: ContentContext,
   width: '100%',
   aspectRatio: 1,
   backgroundColor: '$secondaryBackground',
