+++ packages/shared/src/db/modelBuilders.ts
@@ -112,7 +112,10 @@ export function assemblePostFromActivityEvent(event: db.ActivityEvent) {
     sentAt: event.timestamp,
     receivedAt: event.timestamp,
     content: JSON.stringify(postContent),
-    textContent: ub.getTextContent(event.content as ub.Story),
+    textContent: logic.getTextContent(
+      event.content as ub.Story,
+      logic.PlaintextPreviewConfig.inlineConfig
+    ),
     images: api.getContentImages(event.id, event.content as ub.Story),
     reactions: [],
     replies: [],
@@ -158,7 +161,10 @@ export function buildPendingPost({
     title: metadata?.title ?? '',
     image: metadata?.image ?? '',
     content: JSON.stringify(postContent),
-    textContent: ub.getTextContent(content),
+    textContent: logic.getTextContent(
+      postContent,
+      logic.PlaintextPreviewConfig.inlineConfig
+    ),
     images: api.getContentImages(id, content),
     reactions: [],
     replies: [],
