+++ packages/shared/src/logic/index.ts
@@ -11,4 +11,5 @@ export * from './tiptap';
 export * from './hosting';
 export * from './noun';
 export * from './wayfinding';
+export * from './postContent';
 export * from './content-helpers';
